Add subscription and topic package

+++ emma-common/src/main/java/at/ac/tuwien/dsg/emma/util/Concurrent.java
@@ -12,7 +12,7 @@ public final class Concurrent {
         // util class
     }
 
-    public static void sleep(int ms) {
+    public static void sleep(long ms) {
         try {
             Thread.sleep(ms);
         } catch (InterruptedException e) {
